Rename gotham::protocol to gotham::helpers::http
