+++ examples/example_contribution_template/name/src/main.rs
@@ -25,7 +25,7 @@ extern crate hyper;
 extern crate mime;
 
 use hyper::{Response, StatusCode};
-use gotham::protocol::response::create_response;
+use gotham::helpers::http::response::create_response;
 use gotham::router::Router;
 use gotham::router::builder::*;
 use gotham::state::State;
