+++ examples/headers/setting/src/main.rs
@@ -6,7 +6,7 @@ extern crate hyper;
 extern crate mime;
 
 use hyper::{Response, StatusCode};
-use gotham::protocol::response::create_response;
+use gotham::helpers::http::response::create_response;
 use gotham::router::Router;
 use gotham::router::builder::*;
 use gotham::state::State;
