+++ examples/hello_world/src/main.rs
@@ -6,7 +6,7 @@ extern crate mime;
 
 use hyper::{Response, StatusCode};
 
-use gotham::protocol::response::create_response;
+use gotham::helpers::http::response::create_response;
 use gotham::state::State;
 
 /// Create a `Handler` which is invoked when responding to a `Request`.
