+++ examples/query_string/introduction/src/main.rs
@@ -13,7 +13,7 @@ extern crate serde_json;
 
 use hyper::{Response, StatusCode};
 
-use gotham::protocol::response::create_response;
+use gotham::helpers::http::response::create_response;
 use gotham::router::Router;
 use gotham::router::builder::*;
 use gotham::state::{FromState, State};
