+++ gotham/src/handler/error.rs
@@ -4,7 +4,7 @@ use std::fmt::{self, Debug, Display, Formatter};
 use hyper::{Response, StatusCode};
 
 use handler::IntoResponse;
-use protocol::response::create_response;
+use helpers::http::response::create_response;
 use state::{request_id, State};
 
 /// Describes an error which occurred during handler execution, and allows the creation of a HTTP
