+++ gotham/src/helpers/http/header/x_frame_options.rs
@@ -10,7 +10,7 @@ use hyper;
 use hyper::Uri;
 use hyper::header::{Formatter, Header, Raw};
 
-use protocol::header::from_one_rws_delimited_raw_str;
+use helpers::http::header::from_one_rws_delimited_raw_str;
 
 static NAME: &'static str = "X-Frame-Options";
 
@@ -43,7 +43,7 @@ static NAME: &'static str = "X-Frame-Options";
 /// # extern crate gotham;
 ///
 /// use hyper::header::Headers;
-/// use gotham::protocol::header::XFrameOptions;
+/// use gotham::helpers::http::header::XFrameOptions;
 ///
 /// # fn main () {
 /// let mut headers = Headers::new();
