+++ gotham/src/helpers/http/header/x_request_id.rs
@@ -14,7 +14,7 @@ header! {
     /// # extern crate gotham;
     ///
     /// use hyper::header::Headers;
-    /// use gotham::protocol::header::XRequestId;
+    /// use gotham::helpers::http::header::XRequestId;
     ///
     /// # fn main () {
     /// let mut headers = Headers::new();
