+++ gotham/src/helpers/mod.rs
@@ -1,0 +1,3 @@
+//! Helpers, e.g. for HTTP request handling and response generation
+
+pub mod http;
