+++ gotham/src/router/tree/node.rs
@@ -4,7 +4,7 @@ use std::cmp::Ordering;
 use std::borrow::Borrow;
 use hyper::StatusCode;
 
-use protocol::PercentDecoded;
+use helpers::http::PercentDecoded;
 use router::non_match::RouteNonMatch;
 use router::route::{Delegation, Route};
 use router::tree::{Path, SegmentMapping, SegmentsProcessed};
@@ -381,7 +381,7 @@ mod tests {
     use router::route::matcher::MethodOnlyRouteMatcher;
     use router::route::{Extractors, Route, RouteImpl};
     use pipeline::set::*;
-    use protocol::request::path::RequestPathSegments;
+    use helpers::http::request::path::RequestPathSegments;
     use state::{set_request_id, State};
 
     fn handler(state: State) -> (State, Response) {
