+++ gotham/src/service/timing.rs
@@ -5,7 +5,7 @@ use std::fmt::{self, Display, Formatter};
 use chrono::prelude::*;
 use hyper::Response;
 
-use protocol::header::XRuntimeMicroseconds;
+use helpers::http::header::XRuntimeMicroseconds;
 use state::{request_id, State};
 
 /// Used by `GothamService` to time requests. The `elapsed` function returns the elapsed time
